+++ packages/react/src/components/Chat/ChatMessage.tsx
@@ -18,9 +18,9 @@ import {
   applyAccessibilityKeyHandlers,
 } from '../../lib'
 import { WithAsProp, ShorthandValue, ComponentEventHandler, withSafeTypeForAs } from '../../types'
-import { chatMessageBehavior, toolbarBehavior } from '../../lib/accessibility'
+import { chatMessageBehavior, menuAsToolbarBehavior } from '../../lib/accessibility'
 import { IS_FOCUSABLE_ATTRIBUTE } from '../../lib/accessibility/FocusZone'
-import { Accessibility, AccessibilityActionHandlers } from '../../lib/accessibility/types'
+import { Accessibility } from '../../lib/accessibility/types'
 
 import Box from '../Box/Box'
 import Label from '../Label/Label'
@@ -137,7 +137,7 @@ class ChatMessage extends UIComponent<WithAsProp<ChatMessageProps>, ChatMessageS
     isFromKeyboard: false,
   }
 
-  protected actionHandlers: AccessibilityActionHandlers = {
+  protected actionHandlers = {
     // prevents default FocusZone behavior, e.g., in ChatMessageBehavior, it prevents FocusZone from using arrow keys
     // as navigation (only Tab key should work)
     preventDefault: event => {
@@ -214,7 +214,7 @@ class ChatMessage extends UIComponent<WithAsProp<ChatMessageProps>, ChatMessageS
             {Menu.create(actionMenu, {
               defaultProps: {
                 [IS_FOCUSABLE_ATTRIBUTE]: true,
-                accessibility: toolbarBehavior,
+                accessibility: menuAsToolbarBehavior,
                 className: ChatMessage.slotClassNames.actionMenu,
                 styles: styles.actionMenu,
               },