+++ packages/react/src/index.ts
@@ -150,6 +150,15 @@ export { default as Text } from './components/Text/Text'
 export * from './components/Animation/Animation'
 export { default as Animation } from './components/Animation/Animation'
 
+export * from './components/Toolbar/Toolbar'
+export { default as Toolbar } from './components/Toolbar/Toolbar'
+export * from './components/Toolbar/ToolbarDivider'
+export { default as ToolbarDivider } from './components/Toolbar/ToolbarDivider'
+export * from './components/Toolbar/ToolbarItem'
+export { default as ToolbarItem } from './components/Toolbar/ToolbarItem'
+export * from './components/Toolbar/ToolbarRadioGroup'
+export { default as ToolbarRadioGroup } from './components/Toolbar/ToolbarRadioGroup'
+
 export * from './components/Tree'
 export { default as Tree } from './components/Tree'
 
@@ -171,10 +180,12 @@ export {
 } from './lib/accessibility/Behaviors/Menu/menuDividerBehavior'
 export { default as tabBehavior } from './lib/accessibility/Behaviors/Tab/tabBehavior'
 export { default as tabListBehavior } from './lib/accessibility/Behaviors/Tab/tabListBehavior'
-export { default as toolbarBehavior } from './lib/accessibility/Behaviors/Toolbar/toolbarBehavior'
 export {
-  default as toolbarButtonBehavior,
-} from './lib/accessibility/Behaviors/Toolbar/toolbarButtonBehavior'
+  default as menuAsToolbarBehavior,
+} from './lib/accessibility/Behaviors/Toolbar/menuAsToolbarBehavior'
+export {
+  default as menuItemAsToolbarButtonBehavior,
+} from './lib/accessibility/Behaviors/Toolbar/menuItemAsToolbarButtonBehavior'
 export {
   default as radioGroupBehavior,
 } from './lib/accessibility/Behaviors/Radio/radioGroupBehavior'
@@ -229,3 +240,4 @@ export const FocusZoneUtilities = {
 }
 export * from './lib/accessibility/FocusZone/FocusZone.types'
 export * from './lib/accessibility/types'
+export * from './lib/accessibility/reactTypes'